refactor: split Flask file into individual views, add CSV & Excel sources, apply recommended fixes
