+++ viewer/src/stores/commonStore.js
@@ -7,12 +7,14 @@ const createCommonSlice = (set, get) => {
     const dashboards = project?.project_json?.dashboards ?? [];
     const isNew = project?.project_json?.name === "Quickstart Visivo" && dashboards.length === 0;
     set({ isNewProject: isNew });
+    set({ isOnBoardingLoading: false });
   };
 
   return {
     project: null,
     projectFilePath: null,
     isNewProject: undefined,
+    isOnBoardingLoading: true,
     scrollPositions: {},
     setScrollPosition: (dashName, pos) => {
       set(state => ({
