+++ visivo/models/sources/fields.py
@@ -3,6 +3,8 @@ from typing import Annotated, Any, Union
 from pydantic import Discriminator, Tag
 from visivo.models.base.base_model import ContextStringType, RefStringType
 from visivo.models.base.context_string import CONTEXT_STRING_VALUE_REGEX
+from visivo.models.sources.csv_source import CSVFileSource
+from visivo.models.sources.excel_source import ExcelFileSource
 from visivo.models.sources.mysql_source import MysqlSource
 from visivo.models.sources.postgresql_source import PostgresqlSource
 from visivo.models.sources.snowflake_source import SnowflakeSource
@@ -33,6 +35,8 @@ SourceField = Annotated[
         Annotated[SnowflakeSource, Tag("snowflake")],
         Annotated[BigQuerySource, Tag("bigquery")],
         Annotated[DuckdbSource, Tag("duckdb")],
+        Annotated[CSVFileSource, Tag("csv")],
+        Annotated[ExcelFileSource, Tag("xls")],
     ],
     Discriminator(get_model_discriminator_value),
 ]
@@ -46,6 +50,8 @@ SourceRefField = Annotated[
         Annotated[MysqlSource, Tag("mysql")],
         Annotated[SnowflakeSource, Tag("snowflake")],
         Annotated[BigQuerySource, Tag("bigquery")],
+        Annotated[CSVFileSource, Tag("csv")],
+        Annotated[ExcelFileSource, Tag("xls")],
     ],
     Discriminator(get_model_discriminator_value),
 ]
